Implementacion del caso r de juego, ahora cuando pulsas en los nuevos juegos creados del panel de administrador te lleva a una nueva ventana.

+++ levelupgamesTFG/src/main/java/com/tfg/levelupgames/repositories/PrecioRepository.java
@@ -1,5 +1,6 @@
 package com.tfg.levelupgames.repositories;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,6 +10,6 @@ import com.tfg.levelupgames.entities.Precio;
 
 @Repository
 public interface PrecioRepository extends JpaRepository<Precio, Long> {
-    Optional<Precio> findByCantidad(Double cantidad);
+    Optional<Precio> findByCantidad(BigDecimal cantidad);
     Optional<Precio> findById(int id);
 }
